Initial Commit
Two sum problem: you are given an array of numbers and a sum. you need to print out the 2 elements that add up to that exact sum.
now this can be solved by 2 ways
1) creating 2 for loops and sequentially going through each pair of elements. It is east but not time efficient because of 2 for loops.
2) using the hash table. create an empty list put first the element in it and find out if sum-element is in that list. is it is there then append it to list else discard.
for example lets suppose this is the list (3, 4, 5, 1, 2, 7) and sum is 9
now for first element 9-3=6..six is not in list so discard both
             second element9-4=5, five is in list so append both 4 &5...
this method is time efficient. since there is just 1 loop.
I have tried to write it in as simple and easy code as possible.